Fixing a code smell on TimeInterval

+++ src/main/java/com/clocked/worktimecalculator/entities/TimeInterval.java
@@ -28,7 +28,7 @@ public class TimeInterval {
   }
 
   public double durationInHours() {
-    return durationInSeconds() / (60 * 60);
+    return (double) durationInSeconds() / (60 * 60);
   }
 
   public TimeIntervalType type() {
